Large refactor to use Spring Security.
This has the LTI launch create a spring security context. As this is pretty much a fork from the original kstate lti-launch that was really helpful.

Although it has the artifactId it changes the groupId to uk.ac.ox.it.vle. It leaves the packages as they were from the original fork so it’s easier to see what’s changed.

It removes the OAuth2 code from the lti-launch project as that in newer spring versions can be handled by spring-security-oauth2. 

It switches from using a controller to handle the LTI launch to using filters and these filters are dropped into the Spring Security chain.

It also allows multiple LTI consumers to be registered although the example implementation included just supports a single one.