Add support for keeping the Locale in the session.
The locale in the Session can be either the one passed across on the LTI launch or null to indicate that we don’t know the locale from the LTI launch. It’s then up to the application to decide what the fallback should be.

It could at this point use the browser headers or fallback to the system default.