Just try to stop launch data from being null.
When Spring autocreates an instance of LtiSession we should have an empty launch data rather than a null one. The allows for less null checks in calling code.

It doesn’t stop someone from explicitly setting launch data to null but makes it harder to accidentally end up with it being null.